refactor: structuring folder architecture in a different way and adding @ operator for source imports
