+++ src/index.css
@@ -4,6 +4,7 @@
     font-weight: 400;
 
     background-color: #242424;
+    color-scheme: dark;
     color: rgba(255, 255, 255, 0.87);
 
     font-synthesis: none;
