+++ src/main.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router';
 import App from './App';
 import './index.css';
 import { Provider } from 'react-redux';
-import { store } from './utils/redux/store';
+import { store } from './redux/store';
 
 const root = document.getElementById('root')!;
