+++ src/redux/store/index.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { authSlice } from './auth/authSlice';
+import { authSlice } from '@/redux/slices/authSlice';
 
 export const store = configureStore({
     reducer: {
